Factor out native functions into their own file. Use a macro to check for arugment count in native functions
