+++ src/native.h
@@ -1,0 +1,10 @@
+#ifndef nqq_native_h
+#define nqq_native_h
+
+#include "value.h"
+#include "object.h"
+#include "vm.h"
+
+void defineNatives(VM* vm);
+
+#endif
