+++ src/vm.h
@@ -54,5 +54,6 @@ void freeVM();
 InterpretResult interpret(const char* source);
 void push(Value value);
 Value pop();
+bool isFalsey(Value value);
 
 #endif
