+++ test/builtin/print/missing_argument.nqq
@@ -1,1 +1,1 @@
-print(); // expect runtime error: print expected 1 argument but got 0.
+print(); // expect runtime error: print expected 1 arguments but got 0.
