+++ test/builtin/write/missing_argument.nqq
@@ -1,1 +1,1 @@
-write(); // expect runtime error: write expected 1 argument but got 0.
+write(); // expect runtime error: write expected 1 arguments but got 0.
